+++ admin/tool/mucertify/classes/local/source/base.php
@@ -71,6 +71,15 @@ abstract class base {
     }
 
     /**
+     * Can a new source of this type be added to programs when creating program?
+     *
+     * @return bool
+     */
+    public static function is_new_allowed_in_new(): bool {
+        return false;
+    }
+
+    /**
      * Can existing source of this type be updated or deleted to certifications?
      *
      * NOTE: Existing enabled sources in certifications cannot be deleted/hidden